i2c: add support for Unisoc "hardware channel" I2C
Some Unisoc SoCs such as UMS9230 contain an additional I2C controller
similar to the one supported by the i2c-sprd driver but incompatible
with it. This controller only supports two-byte transfers and is used
by the hardware DVFS module for controlling an external CPU voltage
regulator, but other peripherals such as a battery charger may also be
connected to it.

Add a driver for it based on the existing i2c-sprd driver and code from
Unisoc's downstream Android kernel. Remove runtime PM support since this
is intended to be an always-on peripheral controller.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>