net: wwan: add simple rpmsg-based network driver
The rpmsg framework is often used for communicating with a built-in
modem processor in various mobile SoCs. On many Unisoc SoCs with LTE
modems, rpmsg channels are available that can be used directly for
transmitting and receiving raw IP packets. Add a simple network driver
based on the rpmsg API for such modems.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>