iommu/sprd: add support for reserved regions
Add support for initial mappings defined in the device tree. Since these
are set up before the device is attached, move the page table allocation
to a separate "finalize" function that is invoked as soon as a device is
available.

Initial mappings (e.g. for the boot framebuffer) often require the
virtual address to match the physical address. To allow configuring the
IOMMU's virtual address range, read the new "sprd,iova-range" property
from the device tree.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>