soc: sprd: add AGDSP communication driver
Add an IPC driver for the audio DSP found in the Unisoc UMS9230 SoC. The
DSP uses a simple IPC protocol based on a fixed-size command ring buffer
and needs to be woken up using a mailbox. Since the DSP also manages the
power domain for the entire audio subsystem and needs to be woken up
using the same mailbox channel to power it on, this driver also acts as
a power domain driver for the audio subsystem.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>