ASoC: sprd: mcdt: add support for MCDT revision 2
Add support for the MCDT version found in newer Unisoc SoCs such as
UMS9230. In this version, the register layout has been changed and one
ADC channel has been replaced with a DAC channel.

Also add support for enabling the audio power domain and MCDT clock,
which is required on UMS9230.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>