ASoC: sprd: add support for single-channel DMA with interleaved data
When MCDT is used, there is only one FIFO for both audio channels. Allow
restricting the number of hardware channels in the DMA parameters to
pass interleaved data directly to the destination.

FIXME: although the driver explicitly requests the interleaved format,
the remaining code for handling multiple channels does not seem to work
correctly with interleaved data!

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>