ASoC: Add support for the UMS9230 digital codec
In the UMS9230 SoC, the audio codec is split into a digital and analog
part. The analog part is integrated into the PMIC (SC2730). The digital
part is integrated into the SoC itself and can be configured via MMIO
when the audio subsystem is powered on.

Add a driver for the digital part.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>