media: platform: sprd: Add Unisoc camera subsystem driver
Add a driver for the camera subsystem found in the UMS9230 SoC.
The camera subsystem contains three CSI receivers which can be
independently connected to different CSI PHY ports and three DCAM
blocks which can be connected to the CSI receivers. The third DCAM block
has a reduced feature set and is not supported by this driver at the
moment.

Besides writing raw frames to memory, the full-featured DCAM blocks also
provide some simple ISP features. Black level correction, white balance
gain control, lens shading correction and automatic exposure statistics
are implemented in this driver. There is also a separate ISP used for
debayering, color correction and denoising, which operates as a
memory-to-memory device and is going to be supported by a separate
driver.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>